Fixed sonar warnings.

+++ src.save/main/java/g0001_0100/s0001_two_sum/Solution.java
@@ -10,7 +10,7 @@ public class Solution {
     public int[] twoSum(int[] numbers, int target) {
         Map<Integer, Integer> indexMap = new HashMap<>();
         for (int i = 0; i < numbers.length; i++) {
-            Integer requiredNum = (Integer) (target - numbers[i]);
+            Integer requiredNum = target - numbers[i];
             if (indexMap.containsKey(requiredNum)) {
                 return new int[] {indexMap.get(requiredNum), i};
             }
